[P4PU-660] ordering tests
