+++ tests/user.spec.ts
@@ -27,3 +27,21 @@ test("[E2E-ARC-2] Come Cittadino autenticato voglio accedere alla sezione 'I mie
   await expect(page.getByTestId('app.user.info.identifier.value')).toContainText(user.fiscalCode);
   await expect(page.getByTestId('app.user.info.email.value')).toContainText(user.email);
 });
+
+test("[E2E-ARC-3] Come Cittadino voglio 'sloggarmi' dall'applicativo", async ({ page }) => {
+  // mock BE response to avoid accessToken cancel
+  await page.route('*/**/arc/v1/logout', async (route) => {
+    const json = {};
+    await route.fulfill({ json });
+  });
+
+  await page.goto('/pagamenti/');
+  await expect(page).toHaveURL('/pagamenti/');
+  await expect(page.getByLabel('party-menu-button')).toBeVisible();
+  await page.getByLabel('party-menu-button').click();
+  await page.getByRole('menuitem').getByText('Esci').click();
+  await page.waitForURL('**/login');
+
+  const accessToken = await page.evaluate(() => localStorage.getItem('accessToken'));
+  expect(accessToken).toBeNull();
+});
